+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Added Proxy tile @pnicolli
 - Add toast component @sneridagh
 - Add sidebar support for components @sneridagh